fix: resolve gosec, errorlint and unconvert findings
Fix the issues the new golangci-lint config surfaced:

- read config files via filepath.Clean to satisfy gosec G304
- tighten ENC cache permissions to 0750/0600 (G301/G306)
- use errors.As instead of a type assertion in isNotFound (errorlint)
- give the mock server a ReadHeaderTimeout (G114) and annotate its
  debug log lines, which print CLI flag values and request data by
  design (G706)
- clamp intstrInt before the int32 conversion (G115)
- drop unnecessary conversions around gwapiv1.PortNumber (unconvert)

+++ internal/controller/pool_controller_test.go
@@ -297,7 +297,7 @@ func TestPoolReconcile_TLSRouteCustomPort(t *testing.T) {
 	if err := c.Get(testCtx(), types.NamespacedName{Name: "puppet", Namespace: testNamespace}, route); err != nil {
 		t.Fatalf("TLSRoute not created: %v", err)
 	}
-	if route.Spec.Rules[0].BackendRefs[0].Port == nil || int32(*route.Spec.Rules[0].BackendRefs[0].Port) != 9140 {
+	if route.Spec.Rules[0].BackendRefs[0].Port == nil || *route.Spec.Rules[0].BackendRefs[0].Port != 9140 {
 		t.Errorf("expected port 9140 on backend ref")
 	}
 }
